Use extra key stretching for generated hash
In contrast to for instance servers which need to be able to generate
quite a lot of hashes for multiple users per second, it is not that time
critical for the common web user.

With this patch the final hash consists of 256 rounds of
hmac-sha512 hashing (and base 64 encoding). This takes about 0.7
seconds on a Intel Core 2 Duo T7300 with Firefox 5.0 and should therefore
be mostly unnoticed for the user.

The benefit of this is, that an attacker which knows the realm and the
output hash needs about 256 times longer to crack the passphrase.

Signed-off-by: Linus Lüssing <linus.luessing@web.de>