Make number of hashing rounds configurable via about:config
With this commit, the number of rounds can be set via about:config. The
default stays 256. Users with faster CPUs and higher security needs can
increase this value.

Note that with changing the rounds value, also the password for every
site where pwdhash was used with a different rounds value needs to be
reset.

Signed-off-by: Linus Lüssing <linus.luessing@web.de>